2021/2021-01-29_why-hydra/requirements.txt
hydra-colorlog==1.0.0
hydra-core==1.0.5
pytorch-lightning==1.1.6
